Sprites added! Code adapted show the appropriate sprite.
The sprites for scorp and moby were added. Some code was modified to display the appropriate sprites (more detections in place for which hitbox is being used + what sprite is being drawn). The GamePanel also has a new function to load images into a BufferedImage array when it is constructed.